app-api/pyproject.toml
[tool.poetry]
name = "app-api"
version = "0.1.0"
description = ""
authors = ["Costa <jorgemdcosta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
fastapi = "^0.109.0"
pydantic = "^2.5.3"
uvicorn = "^0.27.0"
requests = "^2.31.0"
sqlalchemy = "^2.0.25"
pydantic-settings = "^2.1.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
